Moved databases viewset to its own file.
